Move speaker action to core & remove  http fix url
