+++ src/core/speakers/speakerActionTypes.js
