+++ src/core/speakers/speakerReducer.js
