+++ src/core/speakers/speakerSelectors.js
