+++ src/feedback/sessions/SessionsItem.js
@@ -6,7 +6,7 @@ import Grid from '@material-ui/core/Grid'
 import Paper from '@material-ui/core/Paper'
 import { Link } from 'react-router-dom'
 import connect from 'react-redux/es/connect/connect'
-import { getSpeakersListSelector } from '../speaker/core/speakerSelectors'
+import { getSpeakersListSelector } from '../../core/speakers/speakerSelectors'
 import SpeakerList from '../speaker/SpeakerList'
 import { getDateFromStartTime } from '../../core/sessions/sessionsUtils'
 import { getProjectSelector } from '../project/projectSelectors'
